requirements.txt
tqdm
matplotlib
numpy
aiohttp
PyYAML